+++ resources/index.js
@@ -1,0 +1,2 @@
+// And now load this bundle's actual entry point.
+import("./cse");